// src/app/core/interface/producto.interface.ts
export interface ProductoInterface {
  imagenPath:File,
  nombre: string;
  sku: string;
  precio: string;
  talla: string;
  descripcion:string;
  createAt: Date;
}

export interface crearProductoInterface {
  imagenPath: File,
  nombre: string;
  sku: string;
  precio: string;
  talla:string;
  descripcion:string;
}